Extract Identity entities
